Move JSVGRasterizer unit test to SVG fragment
The JSVGRasterizerTest is currently implemented in the ordinary SWT test
bundle, even though the JSVGRasterizer is only one possible (and the
current default) SVGRasterizer implementation. The test is actually a
unit test for the provider, thus it is properly placed in the SVG
fragment that provides the rasterizer implementation itself.

This change moves the test class to the SVG fragment and enables it via
pure surefire execution. This is similar to what is already done
regarding execution of OS-specific unit tests in the OS fragments.
